frontends/benchmarks/deepresearch_bench/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools]
packages = ["aiq_drb_evaluator"]
package-dir = {"aiq_drb_evaluator" = "src"}

[project]
name = "aiq-drb-evaluator"
version = "1.0.0"
description = "Deep Research Bench (DRB) evaluators for NAT - RACE and FACT evaluation methodologies"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "nvidia-nat==1.4.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]

[project.entry-points."nat.plugins"]
drb_evaluators = "aiq_drb_evaluator.register"

